Fix socket event 'NEW_MESSAGE' listener logic by checking type
Fix bug in 'NEW_MESSAGE' listener (socket server) in which
direct messages check was faulty (`messages.name` is always falsey
in payload). Thus, check for new argument `type` instead.

The bug used to add a channel room to a user's direct messages
sidebar if messaged in the channel when user isn't connected.