Rename socket event 'FIND_OR_CREATE_ROOM' to 'DIRECT_MESSAGE_ROOM'
The name is more semantic because finding or creating a room is
done when looking to direct message another user (channels, on the
other hand, must be created first).

Also, the payload of the event is changed from an object to an
array of user ids for the room.

The creator id is no longer needed since that can be accessed
from the socket server.

Update and tidy `socket.io/room.js` and <RoomFinder>